fix activity fit bounds with antimeridian
