+++ apps/fishing-map/features/workspace/activity/ActivityFitBounds.tsx
@@ -1,0 +1,35 @@
+import { useTranslation } from 'react-i18next'
+import { StatFields } from 'queries/stats-api'
+import { IconButton } from '@globalfishingwatch/ui-components'
+import { Bbox, wrapBBoxLongitudes } from '@globalfishingwatch/data-transforms'
+import { useMapFitBounds } from 'features/map/map-viewport.hooks'
+import { FIT_BOUNDS_ANALYSIS_PADDING } from 'data/config'
+
+type ActivityFitBoundsProps = {
+  stats: StatFields
+  loading: boolean
+}
+
+function ActivityFitBounds({ stats, loading }: ActivityFitBoundsProps): React.ReactElement {
+  const { t } = useTranslation()
+  const fitMapBounds = useMapFitBounds()
+  const statsBbox =
+    stats && wrapBBoxLongitudes([stats.minLon, stats.minLat, stats.maxLon, stats.maxLat] as Bbox)
+
+  const onFitBoundsHandle = () => {
+    fitMapBounds(statsBbox, { padding: FIT_BOUNDS_ANALYSIS_PADDING })
+  }
+
+  return statsBbox ? (
+    <IconButton
+      icon="target"
+      tooltip={t('layer.activity_fit_bounds', 'Center view on activity')}
+      loading={loading}
+      tooltipPlacement="top"
+      size="small"
+      onClick={onFitBoundsHandle}
+    />
+  ) : null
+}
+
+export default ActivityFitBounds
